# swagger.yaml
swagger: "2.0"
info:
  description: "This is an api for CUPartTime Project"
  version: "1.0.0"
  title: "CUPartTime"
  license:
    name: "ISC"
    url: "https://opensource.org/licenses/ISC"
host: "localhost:9000"
basePath: "/v2"
...
schemes:
- "https"
- "http"

#-------------------------------------------------------------[JAMES]---------------------------------------------------------------------------
paths:
...
  /api/wallet/job/:id:
    post:
      tags:
      - "other"
      summary: "Employer pay for emlpoyees"
      description: ""
      operationId: "makeTransaction"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        404:
          description: "Not found job with this id."
#-------------------------------------------------------------[porche]---------------------------------------------------------------------------
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
...